requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via pyvespa
aiosignal==1.3.2
    # via aiohttp
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
apsw==3.48.0.0
    # via apswutils
apswutils==0.0.2
    # via fastlite
attrs==25.1.0
    # via aiohttp
beautifulsoup4==4.13.0
    # via python-fasthtml
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
    # via penne-pixels (pyproject.toml)
cryptography==44.0.0
    # via pyvespa
docker==7.1.0
    # via pyvespa
fastcore==1.7.29
    # via
    #   apswutils
    #   fastlite
    #   python-fasthtml
    #   pyvespa
fastlite==0.1.1
    # via python-fasthtml
filelock==3.17.0
    # via torch
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.2.0
    # via torch
ftfy==6.3.1
    # via clip
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   python-fasthtml
    #   pyvespa
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via python-fasthtml
jinja2==3.1.5
    # via
    #   pyvespa
    #   torch
lxml==5.3.0
    # via pyvespa
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   penne-pixels (pyproject.toml)
    #   torchvision
oauthlib==3.2.2
    # via python-fasthtml
packaging==24.2
    # via
    #   clip
    #   fastcore
pillow==11.1.0
    # via torchvision
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
python-dateutil==2.9.0.post0
    # via
    #   python-fasthtml
    #   pyvespa
python-dotenv==1.0.1
    # via
    #   penne-pixels (pyproject.toml)
    #   uvicorn
python-fasthtml==0.12.1
    # via penne-pixels (pyproject.toml)
python-multipart==0.0.20
    # via python-fasthtml
pyvespa==0.53.0
    # via penne-pixels (pyproject.toml)
pyyaml==6.0.2
    # via uvicorn
regex==2024.11.6
    # via clip
requests==2.32.3
    # via
    #   docker
    #   pyvespa
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via pyvespa
ruff==0.9.4
    # via penne-pixels (pyproject.toml)
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
starlette==0.45.3
    # via python-fasthtml
sympy==1.13.3
    # via torch
tenacity==9.0.0
    # via pyvespa
torch==2.2.1
    # via
    #   penne-pixels (pyproject.toml)
    #   clip
    #   torchvision
torchvision==0.17.1
    # via clip
tqdm==4.67.1
    # via clip
typing-extensions==4.12.2
    # via
    #   anyio
    #   beautifulsoup4
    #   pyvespa
    #   torch
urllib3==2.3.0
    # via
    #   docker
    #   requests
uvicorn==0.34.0
    # via python-fasthtml
uvloop==0.21.0
    # via uvicorn
vespacli==8.471.25
    # via penne-pixels (pyproject.toml)
watchfiles==1.0.4
    # via uvicorn
wcwidth==0.2.13
    # via ftfy
websockets==14.2
    # via uvicorn
yarl==1.18.3
    # via aiohttp
